Create WeatherForecastRepository class
