+++ src/ApplicationCore/Entities/WeatherForecast.cs
@@ -1,4 +1,4 @@
-using System.ComponentModel.DataAnnotations.Schema;
+﻿using System.ComponentModel.DataAnnotations.Schema;
 
 namespace ApplicationCore.Entities;
