+++ src/Infrastructure/WeatherForecastRepository.cs
@@ -1,0 +1,24 @@
+﻿using ApplicationCore.Entities;
+using ApplicationCore.Interfaces;
+
+namespace Infrastructure;
+
+public class WeatherForecastRepository : IWeatherForecastRepository
+{
+    private static readonly string[] Summaries = new[]
+    {
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
+
+    /// <inheritdoc/>
+    public IEnumerable<WeatherForecast> GetWeatherForecasts()
+    {
+        var entities = Enumerable.Range(1, 5).Select(index => new WeatherForecast
+        {
+            Date = DateTime.Now.AddDays(index),
+            TemperatureC = Random.Shared.Next(-20, 55),
+            Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+        });
+        return entities;
+    }
+}
